data-producer/requirements.txt
requests
kafka-python
sqlalchemy
psycopg2-binary
six
